requirements.txt
discord.py 
beautifulsoup4 
selenium 
supabase 
webdriver-manager